refactor: screen/home import 시 @/components 경로로 통일
