+++ src/components/index.ts
@@ -1,0 +1,3 @@
+export * from './common';
+export * from './feed';
+export * from './onboarding';
