+++ src/components/onboarding/index.ts
@@ -1,0 +1,6 @@
+export { default as ErrorMessage } from './ErrorMessage';
+export { default as Heading } from './Heading';
+export { default as HeadingDescription } from './HeadingDescription';
+export { default as Label } from './Label';
+export { default as MultiSelectItem } from './MultiSelectItem';
+export { default as SelectItem } from './SelectItem';
